PowerPoint/slide_generator/requirements.txt
python-pptx==0.6.21
openai==0.27.0